+++ CHANGELOG.md
@@ -25,6 +25,7 @@ First release with a MVP of herenciapp, with all the customer-kitchen-waiter-che
 ### Changed
 - Removed placeholders for main and changed for real data in `src/templates/main.html`
 - Correct time zone
+- Added a Column to allow datetime receive timezone values in `src/api/v1/apps/orders/models.py`
 
 ### Fixed
 - Added role checker to tickets and meat endpoints